third_party/blink/web_tests/platform/mac-mac10.12/virtual/unified-autoplay/external/wpt/feature-policy/feature-policy-frame-policy-timing.https.sub-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught TypeError: callback is not a function
FAIL allow attr timing test same origin assert_equals: expected "NotAllowedError" but got "timeout"
PASS allow attr timing test diff origin
Harness: the test ran to completion.
